Refactor and enhance modal components for highlights and programs management
- Removed the ViewDetailsModal component and replaced it with HighlightDetailsModal and ProgramDetailsModal in the portal for better modularity.
- Updated styles for modal components to improve layout and responsiveness across different screen sizes.
- Adjusted CSS properties for better visual consistency and user experience in modal displays.
- Enhanced the AdminHighlightsPage and AdminProgramsPage to utilize the new modal components, streamlining the user interface.
