+++ frontend/src/app/(public)/components/GlobalLoginModal/GlobalLoginModal.module.css
@@ -22,7 +22,7 @@
 .modalContent {
   background: white;
   border-radius: 12px;
-  padding: 2rem;
+  padding: 2rem 2rem 1.25rem 2rem;
   max-width: 500px;
   width: 90%;
   max-height: 90vh;
@@ -53,6 +53,7 @@
   justify-content: center;
   border-radius: 50%;
   transition: all 0.2s ease;
+  z-index: 10;
 }
 
 .closeButton:hover {
@@ -63,7 +64,6 @@
 .modalHeader {
   text-align: center;
   margin-bottom: 1.5rem;
-  padding-right: 2rem;
 }
 
 .modalHeader h2 {
@@ -122,7 +122,7 @@
 }
 
 .button {
-  padding: 0.75rem 1.5rem;
+  padding: 0.625rem 1.5rem;
   border: none;
   border-radius: 8px;
   font-weight: 600;
@@ -155,7 +155,7 @@
   background: none;
   color: #686d71;
   border: none;
-  padding: 0.5rem 1rem;
+  padding: 0.4rem 1rem;
   font-weight: 500;
   text-decoration: none;
   text-underline-offset: 2px;
@@ -176,7 +176,7 @@
   }
 
   .modalContent {
-    padding: 2.25rem;
+    padding: 2.25rem 2.25rem 1.5rem 2.25rem;
     margin: 0;
     width: 100%;
     max-width: 100%;
@@ -193,7 +193,6 @@
   }
   
   .modalHeader {
-    padding-right: 2.5rem;
     margin-bottom: 1rem;
   }
   
@@ -231,13 +230,13 @@
   
   .button {
     max-width: 100%;
-    padding: 0.875rem 1.25rem;
+    padding: 0.75rem 1.25rem;
     font-size: 0.9rem;
     min-height: 44px;
   }
 
   .textButton {
-    padding: 0.625rem 1rem;
+    padding: 0.5rem 1rem;
   }
 }
 
@@ -249,7 +248,7 @@
   }
 
   .modalContent {
-    padding: 2rem;
+    padding: 2rem 2rem 1.35rem 2rem;
     border-radius: 16px;
   }
