--- frontend/src/app/admin/highlights/components/ViewDetailsModal/ViewDetailsModal.module.css
@@ -1,255 +1,0 @@
-/* Modal Overlay */
-.overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 10000;
-  padding: 1rem;
-}
-
-/* Modal Container */
-.modal {
-  background: white;
-  border-radius: 16px;
-  width: 100%;
-  max-width: 800px;
-  max-height: 90vh;
-  box-shadow: 0 20px 25px -5px rgba(0, 0, 0, 0.1);
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-}
-
-/* Header */
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1.5rem 2rem;
-  border-bottom: 1px solid #e5e7eb;
-  background: #f9fafb;
-}
-
-.title {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #111827;
-  margin: 0;
-  font-family: var(--font-poppins);
-}
-
-.closeButton {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  background: none;
-  border: none;
-  border-radius: 6px;
-  color: #6b7280;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  font-size: 18px;
-}
-
-.closeButton:hover {
-  background-color: #e5e7eb;
-  color: #374151;
-}
-
-/* Content */
-.content {
-  flex: 1;
-  overflow-y: auto;
-  padding: 2rem;
-}
-
-/* Sections */
-.section {
-  margin-bottom: 2rem;
-}
-
-.section:last-child {
-  margin-bottom: 0;
-}
-
-.sectionTitle {
-  font-size: 1rem;
-  font-weight: 600;
-  color: #374151;
-  margin: 0 0 1rem 0;
-  font-family: var(--font-poppins);
-}
-
-.description {
-  font-size: 14px;
-  color: #6b7280;
-  line-height: 1.6;
-  margin: 0;
-  font-family: var(--font-inter);
-}
-
-/* Image Gallery */
-.imageGallery {
-  margin-bottom: 1.5rem;
-}
-
-.mainImage {
-  width: 100%;
-  height: 300px;
-  border-radius: 8px;
-  overflow: hidden;
-  background-color: #f3f4f6;
-  margin-bottom: 1rem;
-}
-
-.mainImageContent {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.imageThumbnails {
-  display: flex;
-  gap: 0.5rem;
-  overflow-x: auto;
-  padding-bottom: 0.5rem;
-}
-
-.thumbnail {
-  flex-shrink: 0;
-  width: 80px;
-  height: 80px;
-  border-radius: 6px;
-  overflow: hidden;
-  border: 2px solid transparent;
-  cursor: pointer;
-  transition: border-color 0.2s ease;
-  background: none;
-  padding: 0;
-}
-
-.thumbnail:hover {
-  border-color: #d1d5db;
-}
-
-.activeThumbnail {
-  border-color: #06100f;
-}
-
-.thumbnailImage {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-/* Video Gallery */
-.videoGallery {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  margin-bottom: 1.5rem;
-}
-
-.videoItem {
-  width: 100%;
-}
-
-.video {
-  width: 100%;
-  max-height: 300px;
-  border-radius: 8px;
-  background-color: #000;
-}
-
-/* File List */
-.fileList {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-}
-
-.fileItem {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  padding: 0.75rem;
-  background-color: #f9fafb;
-  border-radius: 8px;
-  border: 1px solid #e5e7eb;
-}
-
-.fileIcon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  background-color: #e5e7eb;
-  border-radius: 6px;
-  color: #6b7280;
-  font-size: 18px;
-  flex-shrink: 0;
-}
-
-.fileInfo {
-  flex: 1;
-  min-width: 0;
-}
-
-.fileName {
-  font-size: 14px;
-  font-weight: 500;
-  color: #374151;
-  margin-bottom: 0.25rem;
-  font-family: var(--font-inter);
-  word-break: break-word;
-}
-
-.fileDetails {
-  display: flex;
-  gap: 1rem;
-  font-size: 12px;
-  color: #9ca3af;
-  font-family: var(--font-inter);
-}
-
-.fileType {
-  text-transform: uppercase;
-  letter-spacing: 0.025em;
-}
-
-/* Metadata */
-.metadata {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-}
-
-.metaItem {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 14px;
-  font-family: var(--font-inter);
-}
-
-.metaIcon {
-  color: #9ca3af;
-  font-size: 16px;
-  flex-shrink: 0;
-}
-
-.metaLabel {
-  color: #6b7280;
-  font-weight: 500;
-  min-width: 60px;
-}
-
-.metaValue {
-  color: #374151;
-}
