+++ frontend/src/app/admin/highlights/components/index.js
@@ -1,4 +1,3 @@
 export { default as SearchAndFilterControls } from './SearchAndFilterControls/SearchAndFilterControls';
 export { default as HighlightCard } from './HighlightCard/HighlightCard';
-export { default as ViewDetailsModal } from './ViewDetailsModal/ViewDetailsModal';
 export { default as HighlightForm } from './HighlightForm/HighlightForm';
