--- frontend/src/app/admin/programs/components/ViewDetailsModal/ViewDetailsModal.js
@@ -1,638 +1,0 @@
-'use client';
-
-import React, { useState, useEffect } from 'react';
-import Image from 'next/image';
-import { FaTimes, FaEye, FaUsers, FaExclamationTriangle, FaCheck, FaBuilding, FaClock, FaInfoCircle, FaHistory } from 'react-icons/fa';
-import { FiCalendar } from 'react-icons/fi';
-import { hasActiveCollaborations as checkHasActiveCollaborations, getActiveCollaborators } from '@/utils/shared/collaborationStatusUtils';
-import { getProgramImageUrl, getOrganizationImageUrl } from '@/utils/shared/uploadPaths';
-import { formatProgramDates, formatDateShort, formatDateTime } from '@/utils/shared/dateUtils';
-import { getStatusDisplayText } from '@/utils/shared/collaborationStatusUtils';
-import DOMPurify from 'dompurify';
-import styles from './ViewDetailsModal.module.css';
-
-const ViewDetailsModal = ({ 
-  program, 
-  onClose, 
-  // Collaboration mode props
-  mode = 'view', // 'view' or 'collaboration'
-  collaboration = null
-}) => {
-  // State for tabs
-  const [activeTab, setActiveTab] = useState('details');
-  // State for collaboration mode
-  const [imageError, setImageError] = useState(false);
-
-  // Reset tab when modal opens or program changes
-  useEffect(() => {
-    setActiveTab('details');
-  }, [program?.id, mode]);
-
-  // Safety check for collaboration data in collaboration mode
-  if (mode === 'collaboration' && !collaboration) {
-    return null;
-  }
-
-  // Use collaboration data if in collaboration mode, otherwise use program data
-  const data = mode === 'collaboration' ? collaboration : program;
-
-  const getCategoryLabel = (category) => {
-    const categoryMap = {
-      outreach: 'Outreach',
-      education: 'Education',
-      health: 'Health',
-      environment: 'Environment',
-      community: 'Community Development',
-      youth: 'Youth Programs',
-      women: 'Women Empowerment',
-      elderly: 'Elderly Care',
-      disaster: 'Disaster Relief',
-      other: 'Other'
-    };
-    return categoryMap[category] || category || 'Uncategorized';
-  };
-
-
-  const getStatusColor = (status) => {
-    switch (status) {
-      case 'pending':
-        return '#fef3c7';
-      case 'accepted':
-        return '#d1fae5';
-      case 'declined':
-        return '#fee2e2';
-      case 'approved':
-        return '#d1fae5';
-      default:
-        return '#fef3c7';
-    }
-  };
-
-  const getStatusTextColor = (status) => {
-    switch (status) {
-      case 'pending':
-        return '#92400e';
-      case 'accepted':
-        return '#065f46';
-      case 'declined':
-        return '#991b1b';
-      case 'approved':
-        return '#065f46';
-      default:
-        return '#92400e';
-    }
-  };
-
-  const getStatusIcon = (status) => {
-    switch (status) {
-      case 'pending':
-        return <FaClock />;
-      case 'accepted':
-        return <FaCheck />;
-      case 'declined':
-        return <FaTimes />;
-      case 'approved':
-        return <FaCheck />;
-      default:
-        return <FaClock />;
-    }
-  };
-
-  const getRequestTypeText = (requestType) => {
-    return requestType === 'received' ? 'Collaboration Request Received' : 'Collaboration Request Sent';
-  };
-
-  const getRequestTypeColor = (requestType) => {
-    return requestType === 'received' ? '#3b82f6' : '#8b5cf6';
-  };
-
-  return (
-    <div className={styles.modalOverlay} onClick={onClose}>
-      <div className={styles.modalContent} onClick={(e) => e.stopPropagation()}>
-        <div className={styles.modalHeader}>
-          <div className={styles.titleSection}>
-            <h2 className={styles.modalTitle}>
-              {mode === 'collaboration' ? data.program_title : data.title}
-            </h2>
-            {mode === 'collaboration' && (
-              <div className={styles.requestType} style={{ color: getRequestTypeColor(data.request_type) }}>
-                {getRequestTypeText(data.request_type)}
-              </div>
-            )}
-          </div>
-          <div className={styles.headerActions}>
-            {mode === 'collaboration' && (
-              <div className={styles.statusBadge} style={{ 
-                backgroundColor: getStatusColor(data.status),
-                color: getStatusTextColor(data.status)
-              }}>
-                {getStatusIcon(data.status)}
-                <span>{(data.status || 'pending').charAt(0).toUpperCase() + (data.status || 'pending').slice(1)}</span>
-              </div>
-            )}
-            <button onClick={onClose} className={styles.closeButton}>
-              <FaTimes />
-            </button>
-          </div>
-        </div>
-
-        <div className={styles.modalBody}>
-          {/* Tabs - Only show in view mode */}
-          {mode === 'view' && (
-            <div className={styles.tabsContainer}>
-              <button
-                className={`${styles.tab} ${activeTab === 'details' ? styles.activeTab : ''}`}
-                onClick={() => setActiveTab('details')}
-              >
-                <FaInfoCircle className={styles.tabIcon} />
-                Details
-              </button>
-              <button
-                className={`${styles.tab} ${activeTab === 'activity' ? styles.activeTab : ''}`}
-                onClick={() => setActiveTab('activity')}
-              >
-                <FaHistory className={styles.tabIcon} />
-                Activity
-              </button>
-            </div>
-          )}
-
-          {/* Tab Content */}
-          {activeTab === 'details' && (
-            <div className={styles.contentLayout}>
-            {/* Top Section - Image and Program Info Side by Side */}
-            <div className={styles.topSection}>
-              {/* Left - Program Image */}
-              {(mode === 'collaboration' ? data.program_image : data.image) ? (
-                <div className={styles.imageSection}>
-                  {mode === 'collaboration' ? (
-                    // Collaboration mode - direct image URL
-                    <Image
-                      src={data.program_image}
-                      alt={data.program_title}
-                      className={styles.programImage}
-                      width={600}
-                      height={300}
-                      onError={() => {
-                        setImageError(true);
-                      }}
-                    />
-                  ) : (
-                    // View mode - use utility function
-                    <>
-                      {getProgramImageUrl(data.image) === 'IMAGE_UNAVAILABLE' ? (
-                        <div className={styles.imagePlaceholder}>
-                          <FaExclamationTriangle />
-                          <span>Image unavailable</span>
-                        </div>
-                      ) : (
-                        <Image
-                          src={getProgramImageUrl(data.image)}
-                          alt={data.title}
-                          className={styles.programImage}
-                          width={600}
-                          height={300}
-                          onError={() => {
-                            setImageError(true);
-                          }}
-                        />
-                      )}
-                    </>
-                  )}
-                  {mode === 'collaboration' && imageError && (
-                    <div className={styles.imagePlaceholder}>
-                      <FaExclamationTriangle />
-                      <span>Image unavailable</span>
-                    </div>
-                  )}
-                </div>
-              ) : (
-                <div className={styles.imageSection}>
-                  <div className={styles.imagePlaceholder}>
-                    <FaEye />
-                    <span>No image available</span>
-                  </div>
-                </div>
-              )}
-
-              {/* Right - Program Title, Status, Program Details */}
-              <div className={styles.programInfoSection}>
-                {/* Category at the top */}
-                <div className={styles.categoryBadge}>
-                  {getCategoryLabel(mode === 'collaboration' ? data.program_category : data.category)}
-                </div>
-                <h3 className={styles.programTitle}>
-                  {mode === 'collaboration' ? data.program_title : data.title}
-                </h3>
-                
-
-                {/* Program Details */}
-                <div className={styles.detailsSection}>
-                  <div className={styles.detailsGrid}>
-                    <div className={styles.detailItem}>
-                      <FiCalendar className={styles.detailIcon} />
-                      <span className={styles.detailValue}>
-                        {mode === 'collaboration' ? (
-                          data.event_start_date ? formatDateShort(data.event_start_date) : 'N/A'
-                        ) : (
-                          formatProgramDates(data)
-                        )}
-                      </span>
-                    </div>
-
-                    {/* Program Status - show in both modes */}
-                    {(mode === 'collaboration' ? data.program_status : data.status) && (
-                      <div className={styles.detailItem}>
-                        <span className={styles.detailLabel}>Program Status:</span>
-                        <span className={styles.detailValue}>
-                          {mode === 'collaboration' ? data.program_status : (() => {
-                            // Handle special status display for collaborative programs
-                            // Check if program has any active collaborations
-                            const hasActiveCollaborations = checkHasActiveCollaborations(data);
-                            
-                            if (hasActiveCollaborations) {
-                              return getStatusDisplayText(data.status);
-                            }
-                            return data.status.charAt(0).toUpperCase() + data.status.slice(1);
-                          })()}
-                        </span>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            {/* Description - Full Width Below */}
-            <div className={styles.descriptionSection}>
-              <h4 className={styles.sectionTitle}>Description</h4>
-              <div 
-                className={styles.description}
-                dangerouslySetInnerHTML={{ 
-                  __html: (mode === 'collaboration' ? data.program_description : data.description)
-                    ? DOMPurify.sanitize(mode === 'collaboration' ? data.program_description : data.description)
-                    : '<p>No description provided</p>' 
-                }} 
-              />
-            </div>
-
-            {/* Collaboration Section - only show in view mode */}
-            {mode === 'view' && (() => {
-              // Check if program has any active collaborations
-              const hasActiveCollaborations = checkHasActiveCollaborations(data);
-              
-              const hasCollaborators = hasActiveCollaborations &&
-                                     data.collaborators.some(collab => 
-                                       collab && 
-                                       typeof collab === 'object' && 
-                                       collab.email && 
-                                       collab.email.trim() !== ''
-                                     );
-
-              if (!hasCollaborators) {
-                return null;
-              }
-
-              // Get active collaborators (excludes declined/opted-out ones)
-              const validCollaborators = getActiveCollaborators(data.collaborators).filter(collab => 
-                collab && 
-                typeof collab === 'object' && 
-                collab.email && 
-                collab.email.trim() !== ''
-              );
-
-              if (validCollaborators.length === 0) {
-                return null;
-              }
-
-              return (
-                <div className={styles.collaborationSection}>
-                  <h4 className={styles.sectionTitle}>
-                    <FaUsers className={styles.sectionIcon} />
-                    Collaborators
-                  </h4>
-                  <div className={styles.collaboratorsList}>
-                    {validCollaborators.map((collaborator, index) => (
-                      <div key={collaborator.id || `collab-${index}`} className={styles.collaboratorItem}>
-                        <div className={styles.collaboratorInfo}>
-                          <span className={styles.collaboratorEmail}>{collaborator.email}</span>
-                          {collaborator.organization_acronym && 
-                           collaborator.organization_acronym.trim() !== '' && (
-                            <span className={styles.collaboratorOrg}>
-                              ({collaborator.organization_acronym})
-                            </span>
-                          )}
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-              );
-            })()}
-
-            {/* Organization Information - only show in collaboration mode */}
-            {mode === 'collaboration' && (() => {
-              const orgName = data.request_type === 'received' ? data.inviter_org_name : data.invitee_org_name;
-              const orgAcronym = data.request_type === 'received' ? data.inviter_org_acronym : data.invitee_org_acronym;
-              const orgLogo = data.request_type === 'received' ? data.inviter_org_logo : data.invitee_org_logo;
-              const adminEmail = data.request_type === 'received' ? data.inviter_email : data.invitee_email;
-              const logoUrl = orgLogo ? getOrganizationImageUrl(orgLogo, 'logo') : null;
-              
-              return (
-                <div className={styles.section}>
-                  <h3 className={styles.sectionTitle}>Collaborator</h3>
-                  <div className={styles.organizationInfo}>
-                    <div className={styles.orgCard}>
-                      <div className={styles.orgDetails}>
-                        {/* Always show logo container, with fallback if no logo */}
-                        <div className={styles.orgLogoContainer}>
-                          {logoUrl && logoUrl !== 'ORGANIZATION_LOGO_UNAVAILABLE' && orgLogo ? (
-                            <Image
-                              src={logoUrl}
-                              alt={`${orgName} logo`}
-                              width={48}
-                              height={48}
-                              className={styles.orgLogo}
-                            />
-                          ) : (
-                            <div className={styles.orgLogoPlaceholder}>
-                              <FaBuilding />
-                            </div>
-                          )}
-                        </div>
-                        <div className={styles.orgTextContainer}>
-                          <div className={styles.orgName}>
-                            {orgName} ({orgAcronym})
-                          </div>
-                          <div className={styles.adminEmail}>
-                            {adminEmail}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              );
-            })()}
-
-            {/* Timeline - only show in collaboration mode */}
-            {mode === 'collaboration' && (
-              <div className={styles.section}>
-                <h3 className={styles.sectionTitle}>Timeline</h3>
-                <div className={styles.timeline}>
-                  <div className={styles.timelineItem}>
-                    <div className={styles.timelineIcon}>
-                      <FaClock />
-                    </div>
-                    <div className={styles.timelineContent}>
-                      <div className={styles.timelineTitle}>Request Sent</div>
-                      <div className={styles.timelineDate}>{formatDateShort(data.invited_at)}</div>
-                    </div>
-                  </div>
-                  
-                  {data.responded_at && (
-                    <div className={styles.timelineItem}>
-                      <div className={styles.timelineIcon} style={{ backgroundColor: getStatusColor(data.status) }}>
-                        {getStatusIcon(data.status)}
-                      </div>
-                      <div className={styles.timelineContent}>
-                      <div className={styles.timelineTitle}>
-                        Request {(data.status || 'pending') === 'accepted' ? 'Accepted' : 'Declined'}
-                      </div>
-                        <div className={styles.timelineDate}>{formatDateShort(data.responded_at)}</div>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Superadmin Approval Status */}
-                  {!data.is_approved && data.collaboration_status === 'accepted' && (
-                    <div className={styles.timelineItem}>
-                      <div className={styles.timelineIcon} style={{ backgroundColor: '#e0e7ff', color: '#3730a3' }}>
-                        <FaClock />
-                      </div>
-                      <div className={styles.timelineContent}>
-                        <div className={styles.timelineTitle}>Pending Superadmin Approval</div>
-                        <div className={styles.timelineDate}>Awaiting superadmin review</div>
-                      </div>
-                    </div>
-                  )}
-
-                  {data.is_approved && (
-                    <div className={styles.timelineItem}>
-                      <div className={styles.timelineIcon} style={{ backgroundColor: '#10b981', color: 'white' }}>
-                        <FaCheck />
-                      </div>
-                      <div className={styles.timelineContent}>
-                        <div className={styles.timelineTitle}>Superadmin Approved</div>
-                        <div className={styles.timelineDate}>Program approved and published</div>
-                      </div>
-                    </div>
-                  )}
-
-                  {!data.is_approved && data.collaboration_status === 'declined' && (
-                    <div className={styles.timelineItem}>
-                      <div className={styles.timelineIcon} style={{ backgroundColor: '#fee2e2', color: '#991b1b' }}>
-                        <FaTimes />
-                      </div>
-                      <div className={styles.timelineContent}>
-                        <div className={styles.timelineTitle}>Superadmin Rejected</div>
-                        <div className={styles.timelineDate}>Program rejected by superadmin</div>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </div>
-            )}
-
-            {/* Action Information - only show in collaboration mode */}
-            {mode === 'collaboration' && (data.status === 'pending' || data.status === '' || !data.status) && data.request_type === 'received' && (
-              <div className={styles.section}>
-                <div className={styles.actionInfo}>
-                  <FaInfoCircle className={styles.infoIcon} />
-                  <div className={styles.infoContent}>
-                    <h4>Action Required</h4>
-                    <p>
-                      You have received a collaboration request for this program. 
-                      If you accept, the program will be created once all collaborators accept the collaboration request.
-                      If you decline, the program will not be created if no collaborators accept.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {/* Status Information - only show in collaboration mode */}
-            {mode === 'collaboration' && !data.is_approved && data.collaboration_status === 'accepted' && (
-              <div className={styles.section}>
-                <div className={styles.actionInfo}>
-                  <FaInfoCircle className={styles.infoIcon} />
-                  <div className={styles.infoContent}>
-                    <h4>Pending Superadmin Review</h4>
-                    <p>
-                      This collaborative program has been accepted by all collaborators and has been created successfully. 
-                      The program is now live and visible to the public.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {mode === 'collaboration' && data.is_approved && (
-              <div className={styles.section}>
-                <div className={styles.actionInfo}>
-                  <FaInfoCircle className={styles.infoIcon} />
-                  <div className={styles.infoContent}>
-                    <h4>Program Approved</h4>
-                    <p>
-                      This collaborative program has been approved by the superadmin and is now published. 
-                      All collaborating organizations can now view and manage this program.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {mode === 'collaboration' && !data.is_approved && data.collaboration_status === 'declined' && (
-              <div className={styles.section}>
-                <div className={styles.actionInfo}>
-                  <FaInfoCircle className={styles.infoIcon} />
-                  <div className={styles.infoContent}>
-                    <h4>Program Rejected</h4>
-                    <p>
-                      This collaborative program has been rejected by the superadmin. 
-                      The program will not be published and collaboration on this program has ended.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {/* Additional Images - Full Width Below - only show in view mode */}
-            {mode === 'view' && data.additional_images && data.additional_images.length > 0 && (
-              <div className={styles.additionalImagesSection}>
-                <h4 className={styles.sectionTitle}>Additional Images</h4>
-                <div className={styles.additionalImagesGrid}>
-                  {data.additional_images.map((imagePath, index) => (
-                    <div key={index} className={styles.additionalImageContainer}>
-                      <Image
-                        src={getProgramImageUrl(imagePath, 'additional')}
-                        alt={`Additional ${index + 1}`}
-                        className={styles.additionalImage}
-                        width={200}
-                        height={150}
-                      />
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-          </div>
-          )}
-
-          {/* Activity Tracker Tab - Only show in view mode */}
-          {mode === 'view' && activeTab === 'activity' && (
-            <div className={styles.activityTrackerContent}>
-              <h3 className={styles.activityTrackerTitle}>History</h3>
-              
-              <div className={styles.activityList}>
-                {/* Created Activity */}
-                {data.created_at && (
-                  <div className={styles.activityItem}>
-                    <div className={styles.activityIcon}>
-                      <FaClock />
-                    </div>
-                    <div className={styles.activityContent}>
-                      <div className={styles.activityHeader}>
-                        <span className={styles.activityAction}>Program Created</span>
-                        <span className={styles.activityDate}>
-                          {formatDateTime(data.created_at)}
-                        </span>
-                      </div>
-                      {/* Submitted By Information */}
-                      {data.submitted_by_name && (
-                        <div className={styles.activityDetails}>
-                          <div className={styles.activityDetailRow}>
-                            <span className={styles.activityDetailLabel}>Submitted by:</span>
-                            <span className={styles.activityDetailValue}>
-                              {data.submitted_by_name.trim() || 'Not specified'}
-                              {data.submitted_by_role && data.submitted_by_role.trim() && (
-                                <span className={styles.activityRole}> ({data.submitted_by_role})</span>
-                              )}
-                            </span>
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                )}
-
-                {/* Updated Activity - Only show if program has been updated AND has edited_by data */}
-                {(() => {
-                  // Check if edited_by_name exists and has real data
-                  const hasEditedByData = data.edited_by_name && 
-                    typeof data.edited_by_name === 'string' &&
-                    data.edited_by_name.trim() !== '' &&
-                    data.edited_by_name.trim().toLowerCase() !== 'not specified';
-                  
-                  // Check if program has been updated
-                  const hasBeenUpdated = data.updated_at && 
-                    data.created_at && 
-                    (() => {
-                      const updatedTime = new Date(data.updated_at).getTime();
-                      const createdTime = new Date(data.created_at).getTime();
-                      return (updatedTime - createdTime) >= 1000;
-                    })();
-                  
-                  // Only show if program has been updated AND has edited_by data
-                  if (hasBeenUpdated && hasEditedByData) {
-                    return (
-                      <div className={styles.activityItem}>
-                        <div className={styles.activityIcon}>
-                          <FaClock />
-                        </div>
-                        <div className={styles.activityContent}>
-                          <div className={styles.activityHeader}>
-                            <span className={styles.activityAction}>Program Updated</span>
-                            <span className={styles.activityDate}>
-                              {formatDateTime(data.updated_at)}
-                            </span>
-                          </div>
-                          {/* Updated By Information */}
-                          <div className={styles.activityDetails}>
-                            <div className={styles.activityDetailRow}>
-                              <span className={styles.activityDetailLabel}>Updated by:</span>
-                              <span className={styles.activityDetailValue}>
-                                {data.edited_by_name.trim()}
-                                {data.edited_by_role && 
-                                 typeof data.edited_by_role === 'string' &&
-                                 data.edited_by_role.trim() !== '' && (
-                                  <span className={styles.activityRole}> ({data.edited_by_role.trim()})</span>
-                                )}
-                              </span>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    );
-                  }
-                  return null;
-                })()}
-
-                {/* No activity message */}
-                {!data.created_at && !data.updated_at && (
-                  <div className={styles.noActivity}>
-                    <FaHistory className={styles.noActivityIcon} />
-                    <p>No activity history available</p>
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-
-    </div>
-  );
-};
-
-export default ViewDetailsModal;
