+++ frontend/src/app/superadmin/faqs/components/styles/CreateFAQModal.module.css
@@ -57,7 +57,7 @@
   font-weight: 600;
   color: #000;
   margin: 0;
-  font-family: 'Poppins', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+  font-family: var(--font-poppins), sans-serif;
 }
 
 .closeButton {
