+++ frontend/src/app/superadmin/invites/components/styles/InviteModal.module.css
@@ -48,7 +48,7 @@
   font-size: 22px;
   font-weight: 600;
   color: #000;
-  font-family: 'Poppins', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+  font-family: var(--font-poppins), sans-serif;
 }
 
 .closeButton {
