+++ frontend/src/app/superadmin/programs/featured/featured.module.css
@@ -18,7 +18,7 @@
 .pageTitle {
   font-size: 1.25rem;
   font-weight: 700;
-  font-family: 'Poppins', sans-serif;
+  font-family: var(--font-poppins), sans-serif;
   color: #000;
   margin: 0;
   user-select: none;
