+++ frontend/src/app/superadmin/programs/programs.module.css
@@ -182,7 +182,7 @@
   overflow: hidden;
   box-sizing: border-box;
   width: auto;
-  max-width: 400px;
+  max-width: 500px;
 }
 
 .cardContent {
@@ -274,6 +274,14 @@
 
 .approvedCount {
   color: #1d9782;
+}
+
+.archivedCount {
+  font-size: 11.5px;
+  font-weight: 500;
+  font-family: var(--font-inter), sans-serif;
+  line-height: 1.3;
+  color: #8b8e8d;
 }
 
 /* Custom Dropdown Components - Matching approvals */
@@ -468,7 +476,7 @@ ul.options:focus-visible {
 .featuredSection .sectionTitle {
   font-size: 1.15rem;
   font-weight: 700;
-  font-family: 'Poppins', sans-serif;
+  font-family: var(--font-poppins), sans-serif;
   color: #000;
   margin: 2rem 0 2.4rem;
   text-align: left;
@@ -529,9 +537,9 @@ ul.options:focus-visible {
 .programsSection .sectionTitle {
   font-size: 1.15rem;
   font-weight: 700;
-  font-family: 'Poppins', sans-serif;
+  font-family: var(--font-poppins), sans-serif;
   color: #000;
-  margin: 0.5rem 0 2.4rem;
+  margin: 0.5rem 0 1rem;
   user-select: none;
 }
