+++ frontend/src/app/superadmin/settings/TwoFAModal/TwoFAModal.module.css
@@ -96,7 +96,7 @@
   font-size: 12px;
   color: #64748b;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   line-height: 1.4;
 }
 
@@ -191,14 +191,14 @@
   font-weight: 600;
   color: #000;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
 }
 
 .setupDescription {
   font-size: 14px;
   color: #64748b;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   line-height: 1.5;
   max-width: 400px;
 }
@@ -212,7 +212,7 @@
   font-size: 14px;
   font-weight: 600;
   cursor: pointer;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   transition: all 0.2s ease;
   display: flex;
   align-items: center;
@@ -252,7 +252,7 @@
   font-weight: 600;
   color: #374151;
   margin: 0 0 0.5rem 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
 }
 
 .instructionsList {
@@ -267,7 +267,7 @@
 .instructionsList li {
   font-size: 12px;
   color: #64748b;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   line-height: 1.3;
   padding-left: 1.5rem;
   position: relative;
@@ -311,13 +311,13 @@
   font-size: 11px;
   font-weight: 600;
   color: #374151;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
 }
 
 .accountValue {
   font-size: 12px;
   color: #64748b;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   background: white;
   border: 1px solid #e2e8f0;
   border-radius: 6px;
@@ -347,7 +347,7 @@
   font-weight: 500;
   cursor: pointer;
   transition: all 0.2s ease;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -375,7 +375,7 @@
   font-size: 13px;
   color: #64748b;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   text-align: center;
 }
 
@@ -397,7 +397,7 @@
   font-size: 13px;
   font-weight: 600;
   color: #374151;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
 }
 
 .verificationInput {
@@ -406,7 +406,7 @@
   border-radius: 8px;
   font-size: 16px;
   transition: all 0.2s ease;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   background: white;
   text-align: center;
   letter-spacing: 2px;
@@ -435,7 +435,7 @@
   font-size: 13px;
   font-weight: 600;
   cursor: pointer;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   flex: 1;
   user-select: none;
   transition: all 0.2s ease;
@@ -463,7 +463,7 @@
   font-size: 13px;
   font-weight: 600;
   cursor: pointer;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   flex: 1;
   user-select: none;
   transition: all 0.2s ease;
@@ -511,14 +511,14 @@
   font-weight: 600;
   color: #000;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
 }
 
 .successDescription {
   font-size: 14px;
   color: #64748b;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   line-height: 1.5;
   max-width: 300px;
 }
@@ -551,14 +551,14 @@
   font-weight: 600;
   color: #000;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
 }
 
 .enabledDescription {
   font-size: 14px;
   color: #64748b;
   margin: 0;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   line-height: 1.5;
   max-width: 300px;
 }
@@ -572,7 +572,7 @@
   font-size: 14px;
   font-weight: 600;
   cursor: pointer;
-  font-family: var(--font-poppins), sans-serif;
+  font-family: var(--font-inter), sans-serif;
   transition: all 0.2s ease;
   display: flex;
   align-items: center;
