+++ frontend/src/components/portal/HighlightDetailsModal/index.js
@@ -1,0 +1,2 @@
+export { default } from './HighlightDetailsModal'
+
