+++ frontend/src/components/portal/ProgramDetailsModal/index.js
@@ -1,0 +1,2 @@
+export { default } from './ProgramDetailsModal'
+
