+++ frontend/src/components/portal/index.js
@@ -1,3 +1,5 @@
 // Portal components - shared between admin and superadmin portals
 export { Sidebar, LogoutModalTrigger, superadminNavLinks, adminNavLinks } from './Sidebar';
 export { TopBar, adminBreadcrumbConfig, superadminBreadcrumbConfig } from './TopBar';
+export { default as ProgramDetailsModal } from './ProgramDetailsModal';
+export { default as HighlightDetailsModal } from './HighlightDetailsModal';
